requirements.txt
Sphinx==1.2.2
sphinx_rtd_theme==0.1.6
python-slugify==0.0.7
django-tastypie==0.11.1
Django==1.6.5
simplejson==3.5.2